Move alert logic into a separate module
